(PC-36073) refactor(interactionTag): flatten Tag usage
