+++ src/features/venueMap/components/VenueMapBottomSheet/VenueMapOfferPlaylist.tsx
@@ -2,8 +2,7 @@ import React, { Fragment, useCallback } from 'react'
 import { FlatList } from 'react-native-gesture-handler'
 import styled, { useTheme } from 'styled-components/native'
 
-import { getTagConfig } from 'features/offer/components/InteractionTag/getTagConfig'
-import { InteractionTag } from 'features/offer/components/InteractionTag/InteractionTag'
+import { renderInteractionTag } from 'features/offer/components/InteractionTag/InteractionTag'
 import { OfferTile } from 'features/offer/components/OfferTile/OfferTile'
 import { PlaylistType } from 'features/offer/enums'
 import {
@@ -18,7 +17,7 @@ import { Offer } from 'shared/offer/types'
 import { ButtonTertiaryBlack } from 'ui/components/buttons/ButtonTertiaryBlack'
 import { CustomListRenderItem, Playlist } from 'ui/components/Playlist'
 import { PlainArrowNext } from 'ui/svg/icons/PlainArrowNext'
-import { getSpacing, LENGTH_S, RATIO_HOME_IMAGE } from 'ui/theme'
+import { LENGTH_S, RATIO_HOME_IMAGE, getSpacing } from 'ui/theme'
 
 type VenueMapOfferPlaylistProps = {
   offers: Offer[]
@@ -44,7 +43,7 @@ export const VenueMapOfferPlaylist = ({
 
   const renderItem: CustomListRenderItem<Offer> = useCallback(
     ({ item }) => {
-      const tagConfig = getTagConfig({
+      const tag = renderInteractionTag({
         theme,
         likesCount: item.offer.likes,
         chroniclesCount: item.offer.chroniclesCount,
@@ -71,7 +70,7 @@ export const VenueMapOfferPlaylist = ({
           width={PLAYLIST_ITEM_WIDTH}
           height={PLAYLIST_ITEM_HEIGHT}
           playlistType={playlistType}
-          interactionTag={tagConfig ? <InteractionTag {...tagConfig} /> : undefined}
+          interactionTag={tag}
         />
       )
     },
